Results UI improvements
