+++ composeApp/src/commonMain/composeResources/values-as/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
