+++ composeApp/src/commonMain/composeResources/values-kn/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
