+++ composeApp/src/commonMain/composeResources/values-ny-rMW/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
