+++ composeApp/src/commonMain/composeResources/values-pa-rIN/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
