+++ composeApp/src/commonMain/composeResources/values-pt-rMZ/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
