+++ composeApp/src/commonMain/composeResources/values-zh-rHK/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
