+++ composeApp/src/commonMain/composeResources/values-zu-rZA/.gitignore
@@ -1,0 +1,2 @@
+
+strings-organization.xml
